make q-learning algs single-files
