--- baselines/QLearning/buffers/__init__.py
@@ -1,5 +1,0 @@
-# based on dejax https://github.com/hr0nix/dejax/tree/main
-from .uniform import uniform_replay
-from .clustered import clustered_replay
-
-__version__ = '0.1.3'
