+++ baselines/QLearning/iql.py
@@ -20,18 +20,93 @@ import os
 import jax
 import jax.numpy as jnp
 import numpy as np
+from functools import partial
+from typing import NamedTuple, Dict, Union
+
+import chex
 
 import optax
 import flax.linen as nn
 from flax.linen.initializers import constant, orthogonal
 from flax.training.train_state import TrainState
+import flashbax as fbx
 import wandb
 import hydra
 from omegaconf import OmegaConf
+from safetensors.flax import save_file
+from flax.traverse_util import flatten_dict
 
 from jaxmarl import make
+from jaxmarl.wrappers.baselines import CTRolloutManager
 from jaxmarl.environments.smax import map_name_to_scenario
-from baselines.QLearning.utils import CTRolloutManager, EpsilonGreedy, Transition, UniformBuffer, ScannedRNN, save_params
+
+
+class ScannedRNN(nn.Module):
+
+    @partial(
+        nn.scan,
+        variable_broadcast="params",
+        in_axes=0,
+        out_axes=0,
+        split_rngs={"params": False},
+    )
+    @nn.compact
+    def __call__(self, carry, x):
+        """Applies the module."""
+        rnn_state = carry
+        ins, resets = x
+        hidden_size = ins.shape[-1]
+        rnn_state = jnp.where(
+            resets[:, np.newaxis],
+            self.initialize_carry(hidden_size, *ins.shape[:-1]),
+            rnn_state,
+        )
+        new_rnn_state, y = nn.GRUCell(hidden_size)(rnn_state, ins)
+        return new_rnn_state, y
+
+    @staticmethod
+    def initialize_carry(hidden_size, *batch_size):
+        # Use a dummy key since the default state init fn is just zeros.
+        return nn.GRUCell(hidden_size, parent=None).initialize_carry(
+            jax.random.PRNGKey(0), (*batch_size, hidden_size)
+        )
+
+
+class EpsilonGreedy:
+    """Epsilon Greedy action selection"""
+
+    def __init__(self, start_e: float, end_e: float, duration: int):
+        self.start_e  = start_e
+        self.end_e    = end_e
+        self.duration = duration
+        self.slope    = (end_e - start_e) / duration
+        
+    @partial(jax.jit, static_argnums=0)
+    def get_epsilon(self, t: int):
+        e = self.slope*t + self.start_e
+        return jnp.clip(e, self.end_e)
+    
+    @partial(jax.jit, static_argnums=0)
+    def choose_actions(self, q_vals: dict, t: int, rng: chex.PRNGKey):
+        
+        def explore(q, eps, key):
+            key_a, key_e   = jax.random.split(key, 2) # a key for sampling random actions and one for picking
+            greedy_actions = jnp.argmax(q, axis=-1) # get the greedy actions 
+            random_actions = jax.random.randint(key_a, shape=greedy_actions.shape, minval=0, maxval=q.shape[-1]) # sample random actions
+            pick_random    = jax.random.uniform(key_e, greedy_actions.shape)<eps # pick which actions should be random
+            chosed_actions = jnp.where(pick_random, random_actions, greedy_actions)
+            return chosed_actions
+        
+        eps = self.get_epsilon(t)
+        keys = dict(zip(q_vals.keys(), jax.random.split(rng, len(q_vals)))) # get a key for each agent
+        chosen_actions = jax.tree_map(lambda q, k: explore(q, eps, k), q_vals, keys)
+        return chosen_actions
+
+class Transition(NamedTuple):
+    obs: dict
+    actions: dict
+    rewards: dict
+    dones: dict
 
 
 class AgentRNN(nn.Module):
@@ -82,8 +157,14 @@ def make_train(config, env):
             _env_sample_step, env_state, None, config["NUM_STEPS"]
         )
         sample_traj_unbatched = jax.tree_map(lambda x: x[:, 0], sample_traj) # remove the NUM_ENV dim
-        buffer = UniformBuffer(parallel_envs=config["NUM_ENVS"], batch_size=config["BUFFER_BATCH_SIZE"], max_size=config["BUFFER_SIZE"])
-        buffer_state = buffer.reset(sample_traj_unbatched)
+        buffer = fbx.make_flat_buffer(
+            max_length=config['BUFFER_SIZE'],
+            min_length=config['BUFFER_BATCH_SIZE'],
+            sample_batch_size=config['BUFFER_BATCH_SIZE'],
+            add_sequences=True,
+            add_batch_size=None,
+        )
+        buffer_state = buffer.init(sample_traj_unbatched)
 
         # INIT NETWORK
         agent = AgentRNN(action_dim=wrapped_env.max_action_space, hidden_dim=config["AGENT_HIDDEN_DIM"], init_scale=config['AGENT_INIT_SCALE'])
@@ -211,7 +292,8 @@ def make_train(config, env):
             )
 
             # BUFFER UPDATE: save the collected trajectory in the buffer
-            buffer_state = buffer.add(buffer_state, traj_batch)
+            buffer_traj_batch = jax.tree_util.tree_map(lambda x:jnp.swapaxes(x, 0, 1), traj_batch) # put the batch size (num envs) in first axis
+            buffer_state = buffer.add(buffer_state, buffer_traj_batch)
 
             # LEARN PHASE
             def q_of_action(q, u):
@@ -286,7 +368,7 @@ def make_train(config, env):
 
             # sample a batched trajectory from the buffer and set the time step dim in first axis
             rng, _rng = jax.random.split(rng)
-            _, learn_traj = buffer.sample(buffer_state, _rng) # (batch_size, max_time_steps, ...)
+            learn_traj = buffer.sample(buffer_state, _rng).experience.first # (batch_size, max_time_steps, ...)
             learn_traj = jax.tree_map(lambda x: jnp.swapaxes(x, 0, 1), learn_traj) # (max_time_steps, batch_size, ...)
             
             # for iql the loss must be computed differently with or without parameters sharing
@@ -502,6 +584,11 @@ def main(config):
     
     # save params
     if config['SAVE_PATH'] is not None:
+
+        def save_params(params: Dict, filename: Union[str, os.PathLike]) -> None:
+            flattened_dict = flatten_dict(params, sep=',')
+            save_file(flattened_dict, filename)
+
         model_state = outs['runner_state'][0]
         params = jax.tree_map(lambda x: x[0], model_state.params) # save only params of the firt run
         save_dir = os.path.join(config['SAVE_PATH'], env_name)
