+++ requirements/requirements.txt
@@ -1,6 +1,6 @@
 jax>=0.4.11
-flax>=0.7.2
-chex>=0.1.7
+flax==0.7.4
+chex==0.1.81
 dotmap>=1.3.30
 evosax>=0.1.4
 matplotlib>=3.3.4
@@ -15,3 +15,4 @@ hydra-core>=1.3.2
 omegaconf>=2.3.0
 safetensors>=0.3.3
 optax>=0.1.4
+flashbax>=0.0.1
